third_party/WebKit/LayoutTests/external/wpt/html/dom/reflection-metadata-expected.txt

This is a testharness.js-based test.
PASS head.title: 32 tests
PASS head.lang: 32 tests
PASS head.dir: 62 tests
PASS head.className (<head class>): 32 tests
PASS head.hidden: 33 tests
PASS head.accessKey: 32 tests
PASS head.tabIndex: 24 tests
PASS title.title: 32 tests
PASS title.lang: 32 tests
PASS title.dir: 62 tests
PASS title.className (<title class>): 32 tests
PASS title.hidden: 33 tests
PASS title.accessKey: 32 tests
PASS title.tabIndex: 24 tests
PASS base.title: 32 tests
PASS base.lang: 32 tests
PASS base.dir: 62 tests
PASS base.className (<base class>): 32 tests
PASS base.hidden: 33 tests
PASS base.accessKey: 32 tests
PASS base.tabIndex: 24 tests
PASS base.target: 32 tests
PASS link.title: 32 tests
PASS link.lang: 32 tests
PASS link.dir: 62 tests
PASS link.className (<link class>): 32 tests
PASS link.hidden: 33 tests
PASS link.accessKey: 32 tests
PASS link.tabIndex: 24 tests
PASS link.href: 38 tests
PASS link.crossOrigin: 52 tests
PASS link.rel: 32 tests
PASS link.as: 3 tests
FAIL link.as: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected "" but got " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo "
FAIL link.as: setAttribute() to undefined assert_equals: IDL get expected "" but got "undefined"
FAIL link.as: setAttribute() to 7 assert_equals: IDL get expected "" but got "7"
FAIL link.as: setAttribute() to 1.5 assert_equals: IDL get expected "" but got "1.5"
FAIL link.as: setAttribute() to true assert_equals: IDL get expected "" but got "true"
FAIL link.as: setAttribute() to false assert_equals: IDL get expected "" but got "false"
FAIL link.as: setAttribute() to object "[object Object]" assert_equals: IDL get expected "" but got "[object Object]"
FAIL link.as: setAttribute() to NaN assert_equals: IDL get expected "" but got "NaN"
FAIL link.as: setAttribute() to Infinity assert_equals: IDL get expected "" but got "Infinity"
FAIL link.as: setAttribute() to -Infinity assert_equals: IDL get expected "" but got "-Infinity"
FAIL link.as: setAttribute() to "\0" assert_equals: IDL get expected "" but got "\0"
FAIL link.as: setAttribute() to null assert_equals: IDL get expected "" but got "null"
FAIL link.as: setAttribute() to object "test-toString" assert_equals: IDL get expected "" but got "test-toString"
FAIL link.as: setAttribute() to object "test-valueOf" assert_equals: IDL get expected "" but got "test-valueOf"
PASS link.as: setAttribute() to "audio" 
FAIL link.as: setAttribute() to "xaudio" assert_equals: IDL get expected "" but got "xaudio"
FAIL link.as: setAttribute() to "audio\0" assert_equals: IDL get expected "" but got "audio\0"
FAIL link.as: setAttribute() to "udio" assert_equals: IDL get expected "" but got "udio"
FAIL link.as: setAttribute() to "AUDIO" assert_equals: IDL get expected "audio" but got "AUDIO"
PASS link.as: setAttribute() to "document" 
FAIL link.as: setAttribute() to "xdocument" assert_equals: IDL get expected "" but got "xdocument"
FAIL link.as: setAttribute() to "document\0" assert_equals: IDL get expected "" but got "document\0"
FAIL link.as: setAttribute() to "ocument" assert_equals: IDL get expected "" but got "ocument"
FAIL link.as: setAttribute() to "DOCUMENT" assert_equals: IDL get expected "document" but got "DOCUMENT"
PASS link.as: setAttribute() to "embed" 
FAIL link.as: setAttribute() to "xembed" assert_equals: IDL get expected "" but got "xembed"
FAIL link.as: setAttribute() to "embed\0" assert_equals: IDL get expected "" but got "embed\0"
FAIL link.as: setAttribute() to "mbed" assert_equals: IDL get expected "" but got "mbed"
FAIL link.as: setAttribute() to "EMBED" assert_equals: IDL get expected "embed" but got "EMBED"
PASS link.as: setAttribute() to "font" 
FAIL link.as: setAttribute() to "xfont" assert_equals: IDL get expected "" but got "xfont"
FAIL link.as: setAttribute() to "font\0" assert_equals: IDL get expected "" but got "font\0"
FAIL link.as: setAttribute() to "ont" assert_equals: IDL get expected "" but got "ont"
FAIL link.as: setAttribute() to "FONT" assert_equals: IDL get expected "font" but got "FONT"
PASS link.as: setAttribute() to "image" 
FAIL link.as: setAttribute() to "ximage" assert_equals: IDL get expected "" but got "ximage"
FAIL link.as: setAttribute() to "image\0" assert_equals: IDL get expected "" but got "image\0"
FAIL link.as: setAttribute() to "mage" assert_equals: IDL get expected "" but got "mage"
FAIL link.as: setAttribute() to "IMAGE" assert_equals: IDL get expected "image" but got "IMAGE"
PASS link.as: setAttribute() to "manifest" 
FAIL link.as: setAttribute() to "xmanifest" assert_equals: IDL get expected "" but got "xmanifest"
FAIL link.as: setAttribute() to "manifest\0" assert_equals: IDL get expected "" but got "manifest\0"
FAIL link.as: setAttribute() to "anifest" assert_equals: IDL get expected "" but got "anifest"
FAIL link.as: setAttribute() to "MANIFEST" assert_equals: IDL get expected "manifest" but got "MANIFEST"
PASS link.as: setAttribute() to "object" 
FAIL link.as: setAttribute() to "xobject" assert_equals: IDL get expected "" but got "xobject"
FAIL link.as: setAttribute() to "object\0" assert_equals: IDL get expected "" but got "object\0"
FAIL link.as: setAttribute() to "bject" assert_equals: IDL get expected "" but got "bject"
FAIL link.as: setAttribute() to "OBJECT" assert_equals: IDL get expected "object" but got "OBJECT"
PASS link.as: setAttribute() to "report" 
FAIL link.as: setAttribute() to "xreport" assert_equals: IDL get expected "" but got "xreport"
FAIL link.as: setAttribute() to "report\0" assert_equals: IDL get expected "" but got "report\0"
FAIL link.as: setAttribute() to "eport" assert_equals: IDL get expected "" but got "eport"
FAIL link.as: setAttribute() to "REPORT" assert_equals: IDL get expected "report" but got "REPORT"
PASS link.as: setAttribute() to "script" 
FAIL link.as: setAttribute() to "xscript" assert_equals: IDL get expected "" but got "xscript"
FAIL link.as: setAttribute() to "script\0" assert_equals: IDL get expected "" but got "script\0"
FAIL link.as: setAttribute() to "cript" assert_equals: IDL get expected "" but got "cript"
FAIL link.as: setAttribute() to "SCRIPT" assert_equals: IDL get expected "script" but got "SCRIPT"
PASS link.as: setAttribute() to "serviceworker" 
FAIL link.as: setAttribute() to "xserviceworker" assert_equals: IDL get expected "" but got "xserviceworker"
FAIL link.as: setAttribute() to "serviceworker\0" assert_equals: IDL get expected "" but got "serviceworker\0"
FAIL link.as: setAttribute() to "erviceworker" assert_equals: IDL get expected "" but got "erviceworker"
FAIL link.as: setAttribute() to "SERVICEWORKER" assert_equals: IDL get expected "serviceworker" but got "SERVICEWORKER"
PASS link.as: setAttribute() to "sharedworker" 
FAIL link.as: setAttribute() to "xsharedworker" assert_equals: IDL get expected "" but got "xsharedworker"
FAIL link.as: setAttribute() to "sharedworker\0" assert_equals: IDL get expected "" but got "sharedworker\0"
FAIL link.as: setAttribute() to "haredworker" assert_equals: IDL get expected "" but got "haredworker"
FAIL link.as: setAttribute() to "SHAREDWORKER" assert_equals: IDL get expected "sharedworker" but got "SHAREDWORKER"
PASS link.as: setAttribute() to "style" 
FAIL link.as: setAttribute() to "xstyle" assert_equals: IDL get expected "" but got "xstyle"
FAIL link.as: setAttribute() to "style\0" assert_equals: IDL get expected "" but got "style\0"
FAIL link.as: setAttribute() to "tyle" assert_equals: IDL get expected "" but got "tyle"
FAIL link.as: setAttribute() to "STYLE" assert_equals: IDL get expected "style" but got "STYLE"
PASS link.as: setAttribute() to "track" 
FAIL link.as: setAttribute() to "xtrack" assert_equals: IDL get expected "" but got "xtrack"
FAIL link.as: setAttribute() to "track\0" assert_equals: IDL get expected "" but got "track\0"
FAIL link.as: setAttribute() to "rack" assert_equals: IDL get expected "" but got "rack"
FAIL link.as: setAttribute() to "TRACK" assert_equals: IDL get expected "track" but got "TRACK"
PASS link.as: setAttribute() to "video" 
FAIL link.as: setAttribute() to "xvideo" assert_equals: IDL get expected "" but got "xvideo"
FAIL link.as: setAttribute() to "video\0" assert_equals: IDL get expected "" but got "video\0"
FAIL link.as: setAttribute() to "ideo" assert_equals: IDL get expected "" but got "ideo"
FAIL link.as: setAttribute() to "VIDEO" assert_equals: IDL get expected "video" but got "VIDEO"
PASS link.as: setAttribute() to "worker" 
FAIL link.as: setAttribute() to "xworker" assert_equals: IDL get expected "" but got "xworker"
FAIL link.as: setAttribute() to "worker\0" assert_equals: IDL get expected "" but got "worker\0"
FAIL link.as: setAttribute() to "orker" assert_equals: IDL get expected "" but got "orker"
FAIL link.as: setAttribute() to "WORKER" assert_equals: IDL get expected "worker" but got "WORKER"
PASS link.as: setAttribute() to "xslt" 
FAIL link.as: setAttribute() to "xxslt" assert_equals: IDL get expected "" but got "xxslt"
FAIL link.as: setAttribute() to "xslt\0" assert_equals: IDL get expected "" but got "xslt\0"
FAIL link.as: setAttribute() to "slt" assert_equals: IDL get expected "" but got "slt"
FAIL link.as: setAttribute() to "XSLT" assert_equals: IDL get expected "xslt" but got "XSLT"
PASS link.as: IDL set to "" 
FAIL link.as: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected "" but got " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo "
FAIL link.as: IDL set to undefined assert_equals: IDL get expected "" but got "undefined"
FAIL link.as: IDL set to 7 assert_equals: IDL get expected "" but got "7"
FAIL link.as: IDL set to 1.5 assert_equals: IDL get expected "" but got "1.5"
FAIL link.as: IDL set to true assert_equals: IDL get expected "" but got "true"
FAIL link.as: IDL set to false assert_equals: IDL get expected "" but got "false"
FAIL link.as: IDL set to object "[object Object]" assert_equals: IDL get expected "" but got "[object Object]"
FAIL link.as: IDL set to NaN assert_equals: IDL get expected "" but got "NaN"
FAIL link.as: IDL set to Infinity assert_equals: IDL get expected "" but got "Infinity"
FAIL link.as: IDL set to -Infinity assert_equals: IDL get expected "" but got "-Infinity"
FAIL link.as: IDL set to "\0" assert_equals: IDL get expected "" but got "\0"
FAIL link.as: IDL set to null assert_equals: IDL get expected "" but got "null"
FAIL link.as: IDL set to object "test-toString" assert_equals: IDL get expected "" but got "test-toString"
FAIL link.as: IDL set to object "test-valueOf" assert_equals: IDL get expected "" but got "test-valueOf"
PASS link.as: IDL set to "audio" 
FAIL link.as: IDL set to "xaudio" assert_equals: IDL get expected "" but got "xaudio"
FAIL link.as: IDL set to "audio\0" assert_equals: IDL get expected "" but got "audio\0"
FAIL link.as: IDL set to "udio" assert_equals: IDL get expected "" but got "udio"
FAIL link.as: IDL set to "AUDIO" assert_equals: IDL get expected "audio" but got "AUDIO"
PASS link.as: IDL set to "document" 
FAIL link.as: IDL set to "xdocument" assert_equals: IDL get expected "" but got "xdocument"
FAIL link.as: IDL set to "document\0" assert_equals: IDL get expected "" but got "document\0"
FAIL link.as: IDL set to "ocument" assert_equals: IDL get expected "" but got "ocument"
FAIL link.as: IDL set to "DOCUMENT" assert_equals: IDL get expected "document" but got "DOCUMENT"
PASS link.as: IDL set to "embed" 
FAIL link.as: IDL set to "xembed" assert_equals: IDL get expected "" but got "xembed"
FAIL link.as: IDL set to "embed\0" assert_equals: IDL get expected "" but got "embed\0"
FAIL link.as: IDL set to "mbed" assert_equals: IDL get expected "" but got "mbed"
FAIL link.as: IDL set to "EMBED" assert_equals: IDL get expected "embed" but got "EMBED"
PASS link.as: IDL set to "font" 
FAIL link.as: IDL set to "xfont" assert_equals: IDL get expected "" but got "xfont"
FAIL link.as: IDL set to "font\0" assert_equals: IDL get expected "" but got "font\0"
FAIL link.as: IDL set to "ont" assert_equals: IDL get expected "" but got "ont"
FAIL link.as: IDL set to "FONT" assert_equals: IDL get expected "font" but got "FONT"
PASS link.as: IDL set to "image" 
FAIL link.as: IDL set to "ximage" assert_equals: IDL get expected "" but got "ximage"
FAIL link.as: IDL set to "image\0" assert_equals: IDL get expected "" but got "image\0"
FAIL link.as: IDL set to "mage" assert_equals: IDL get expected "" but got "mage"
FAIL link.as: IDL set to "IMAGE" assert_equals: IDL get expected "image" but got "IMAGE"
PASS link.as: IDL set to "manifest" 
FAIL link.as: IDL set to "xmanifest" assert_equals: IDL get expected "" but got "xmanifest"
FAIL link.as: IDL set to "manifest\0" assert_equals: IDL get expected "" but got "manifest\0"
FAIL link.as: IDL set to "anifest" assert_equals: IDL get expected "" but got "anifest"
FAIL link.as: IDL set to "MANIFEST" assert_equals: IDL get expected "manifest" but got "MANIFEST"
PASS link.as: IDL set to "object" 
FAIL link.as: IDL set to "xobject" assert_equals: IDL get expected "" but got "xobject"
FAIL link.as: IDL set to "object\0" assert_equals: IDL get expected "" but got "object\0"
FAIL link.as: IDL set to "bject" assert_equals: IDL get expected "" but got "bject"
FAIL link.as: IDL set to "OBJECT" assert_equals: IDL get expected "object" but got "OBJECT"
PASS link.as: IDL set to "report" 
FAIL link.as: IDL set to "xreport" assert_equals: IDL get expected "" but got "xreport"
FAIL link.as: IDL set to "report\0" assert_equals: IDL get expected "" but got "report\0"
FAIL link.as: IDL set to "eport" assert_equals: IDL get expected "" but got "eport"
FAIL link.as: IDL set to "REPORT" assert_equals: IDL get expected "report" but got "REPORT"
PASS link.as: IDL set to "script" 
FAIL link.as: IDL set to "xscript" assert_equals: IDL get expected "" but got "xscript"
FAIL link.as: IDL set to "script\0" assert_equals: IDL get expected "" but got "script\0"
FAIL link.as: IDL set to "cript" assert_equals: IDL get expected "" but got "cript"
FAIL link.as: IDL set to "SCRIPT" assert_equals: IDL get expected "script" but got "SCRIPT"
PASS link.as: IDL set to "serviceworker" 
FAIL link.as: IDL set to "xserviceworker" assert_equals: IDL get expected "" but got "xserviceworker"
FAIL link.as: IDL set to "serviceworker\0" assert_equals: IDL get expected "" but got "serviceworker\0"
FAIL link.as: IDL set to "erviceworker" assert_equals: IDL get expected "" but got "erviceworker"
FAIL link.as: IDL set to "SERVICEWORKER" assert_equals: IDL get expected "serviceworker" but got "SERVICEWORKER"
PASS link.as: IDL set to "sharedworker" 
FAIL link.as: IDL set to "xsharedworker" assert_equals: IDL get expected "" but got "xsharedworker"
FAIL link.as: IDL set to "sharedworker\0" assert_equals: IDL get expected "" but got "sharedworker\0"
FAIL link.as: IDL set to "haredworker" assert_equals: IDL get expected "" but got "haredworker"
FAIL link.as: IDL set to "SHAREDWORKER" assert_equals: IDL get expected "sharedworker" but got "SHAREDWORKER"
PASS link.as: IDL set to "style" 
FAIL link.as: IDL set to "xstyle" assert_equals: IDL get expected "" but got "xstyle"
FAIL link.as: IDL set to "style\0" assert_equals: IDL get expected "" but got "style\0"
FAIL link.as: IDL set to "tyle" assert_equals: IDL get expected "" but got "tyle"
FAIL link.as: IDL set to "STYLE" assert_equals: IDL get expected "style" but got "STYLE"
PASS link.as: IDL set to "track" 
FAIL link.as: IDL set to "xtrack" assert_equals: IDL get expected "" but got "xtrack"
FAIL link.as: IDL set to "track\0" assert_equals: IDL get expected "" but got "track\0"
FAIL link.as: IDL set to "rack" assert_equals: IDL get expected "" but got "rack"
FAIL link.as: IDL set to "TRACK" assert_equals: IDL get expected "track" but got "TRACK"
PASS link.as: IDL set to "video" 
FAIL link.as: IDL set to "xvideo" assert_equals: IDL get expected "" but got "xvideo"
FAIL link.as: IDL set to "video\0" assert_equals: IDL get expected "" but got "video\0"
FAIL link.as: IDL set to "ideo" assert_equals: IDL get expected "" but got "ideo"
FAIL link.as: IDL set to "VIDEO" assert_equals: IDL get expected "video" but got "VIDEO"
PASS link.as: IDL set to "worker" 
FAIL link.as: IDL set to "xworker" assert_equals: IDL get expected "" but got "xworker"
FAIL link.as: IDL set to "worker\0" assert_equals: IDL get expected "" but got "worker\0"
FAIL link.as: IDL set to "orker" assert_equals: IDL get expected "" but got "orker"
FAIL link.as: IDL set to "WORKER" assert_equals: IDL get expected "worker" but got "WORKER"
PASS link.as: IDL set to "xslt" 
FAIL link.as: IDL set to "xxslt" assert_equals: IDL get expected "" but got "xxslt"
FAIL link.as: IDL set to "xslt\0" assert_equals: IDL get expected "" but got "xslt\0"
FAIL link.as: IDL set to "slt" assert_equals: IDL get expected "" but got "slt"
FAIL link.as: IDL set to "XSLT" assert_equals: IDL get expected "xslt" but got "XSLT"
PASS link.media: 32 tests
FAIL link.nonce: typeof IDL attribute assert_equals: expected "string" but got "undefined"
FAIL link.nonce: IDL get with DOM attribute unset assert_equals: expected (string) "" but got (undefined) undefined
FAIL link.nonce: setAttribute() to "" assert_equals: IDL get expected (string) "" but got (undefined) undefined
FAIL link.nonce: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected (string) " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got (undefined) undefined
FAIL link.nonce: setAttribute() to undefined assert_equals: IDL get expected (string) "undefined" but got (undefined) undefined
FAIL link.nonce: setAttribute() to 7 assert_equals: IDL get expected (string) "7" but got (undefined) undefined
FAIL link.nonce: setAttribute() to 1.5 assert_equals: IDL get expected (string) "1.5" but got (undefined) undefined
FAIL link.nonce: setAttribute() to true assert_equals: IDL get expected (string) "true" but got (undefined) undefined
FAIL link.nonce: setAttribute() to false assert_equals: IDL get expected (string) "false" but got (undefined) undefined
FAIL link.nonce: setAttribute() to object "[object Object]" assert_equals: IDL get expected (string) "[object Object]" but got (undefined) undefined
FAIL link.nonce: setAttribute() to NaN assert_equals: IDL get expected (string) "NaN" but got (undefined) undefined
FAIL link.nonce: setAttribute() to Infinity assert_equals: IDL get expected (string) "Infinity" but got (undefined) undefined
FAIL link.nonce: setAttribute() to -Infinity assert_equals: IDL get expected (string) "-Infinity" but got (undefined) undefined
FAIL link.nonce: setAttribute() to "\0" assert_equals: IDL get expected (string) "\0" but got (undefined) undefined
FAIL link.nonce: setAttribute() to null assert_equals: IDL get expected (string) "null" but got (undefined) undefined
FAIL link.nonce: setAttribute() to object "test-toString" assert_equals: IDL get expected (string) "test-toString" but got (undefined) undefined
FAIL link.nonce: setAttribute() to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (undefined) undefined
FAIL link.nonce: IDL set to "" assert_equals: getAttribute() expected "" but got "test-valueOf"
FAIL link.nonce: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: getAttribute() expected " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got "test-valueOf"
FAIL link.nonce: IDL set to undefined assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
FAIL link.nonce: IDL set to 7 assert_equals: getAttribute() expected "7" but got "test-valueOf"
FAIL link.nonce: IDL set to 1.5 assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
FAIL link.nonce: IDL set to true assert_equals: getAttribute() expected "true" but got "test-valueOf"
FAIL link.nonce: IDL set to false assert_equals: getAttribute() expected "false" but got "test-valueOf"
FAIL link.nonce: IDL set to object "[object Object]" assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
FAIL link.nonce: IDL set to NaN assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
FAIL link.nonce: IDL set to Infinity assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
FAIL link.nonce: IDL set to -Infinity assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
FAIL link.nonce: IDL set to "\0" assert_equals: getAttribute() expected "\0" but got "test-valueOf"
FAIL link.nonce: IDL set to null assert_equals: getAttribute() expected "null" but got "test-valueOf"
FAIL link.nonce: IDL set to object "test-toString" assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
FAIL link.nonce: IDL set to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (object) object "test-valueOf"
PASS link.integrity: 32 tests
PASS link.hreflang: 32 tests
PASS link.type: 32 tests
PASS link.referrerPolicy: 27 tests
FAIL link.referrerPolicy: setAttribute() to "same-origin" assert_equals: IDL get expected "same-origin" but got ""
PASS link.referrerPolicy: 3 tests
FAIL link.referrerPolicy: setAttribute() to "SAME-ORIGIN" assert_equals: IDL get expected "same-origin" but got ""
PASS link.referrerPolicy: 5 tests
FAIL link.referrerPolicy: setAttribute() to "strict-origin" assert_equals: IDL get expected "strict-origin" but got ""
PASS link.referrerPolicy: 3 tests
FAIL link.referrerPolicy: setAttribute() to "STRICT-ORIGIN" assert_equals: IDL get expected "strict-origin" but got ""
PASS link.referrerPolicy: 5 tests
FAIL link.referrerPolicy: setAttribute() to "strict-origin-when-cross-origin" assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
PASS link.referrerPolicy: 3 tests
FAIL link.referrerPolicy: setAttribute() to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN" assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
PASS link.referrerPolicy: 30 tests
FAIL link.referrerPolicy: IDL set to "same-origin" assert_equals: IDL get expected "same-origin" but got ""
PASS link.referrerPolicy: 3 tests
FAIL link.referrerPolicy: IDL set to "SAME-ORIGIN" assert_equals: IDL get expected "same-origin" but got ""
PASS link.referrerPolicy: 5 tests
FAIL link.referrerPolicy: IDL set to "strict-origin" assert_equals: IDL get expected "strict-origin" but got ""
PASS link.referrerPolicy: 3 tests
FAIL link.referrerPolicy: IDL set to "STRICT-ORIGIN" assert_equals: IDL get expected "strict-origin" but got ""
PASS link.referrerPolicy: 5 tests
FAIL link.referrerPolicy: IDL set to "strict-origin-when-cross-origin" assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
PASS link.referrerPolicy: 3 tests
FAIL link.referrerPolicy: IDL set to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN" assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
PASS link.referrerPolicy: 5 tests
PASS link.charset: 32 tests
PASS link.rev: 32 tests
PASS link.target: 32 tests
PASS meta.title: 32 tests
PASS meta.lang: 32 tests
PASS meta.dir: 62 tests
PASS meta.className (<meta class>): 32 tests
PASS meta.hidden: 33 tests
PASS meta.accessKey: 32 tests
PASS meta.tabIndex: 24 tests
PASS meta.name: 32 tests
PASS meta.httpEquiv (<meta http-equiv>): 32 tests
PASS meta.content: 32 tests
PASS meta.scheme: 32 tests
PASS style.title: 32 tests
PASS style.lang: 32 tests
PASS style.dir: 62 tests
PASS style.className (<style class>): 32 tests
PASS style.hidden: 33 tests
PASS style.accessKey: 32 tests
PASS style.tabIndex: 24 tests
PASS style.media: 32 tests
PASS style.type: 32 tests
Harness: the test ran to completion.
